+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Bug fixes
 
 -   Replace PersistentVolume if volume source changes. (https://github.com/pulumi/pulumi-kubernetes/pull/1015).
+-   Fix bool Python provider opts. (https://github.com/pulumi/pulumi-kubernetes/pull/1027).
 
 ## 1.5.6 (February 28, 2020)